[A2Y-249] Address ViewHolder Package
