+++ domain/src/main/java/com/yapp/itemfinder/domain/model/AddItemMarkerMap.kt
@@ -1,7 +1,8 @@
 package com.yapp.itemfinder.domain.model
 
 data class AddItemMarkerMap(
-    override var id: Long = 99,
+    override val id: Long = 99,
+    val lockerEntity: LockerEntity,
     val item: Item? = null,
     override var type: CellType = CellType.ADD_ITEM_MARKER_MAP_CELL
 ): Data(id, type)
