+++ feature/common/src/main/java/com/yapp/itemfinder/feature/common/datalist/adapter/viewholder/AddItemAdditionalViewHolder.kt
@@ -1,7 +1,8 @@
-package com.yapp.itemfinder.feature.common.datalist.adapter
+package com.yapp.itemfinder.feature.common.datalist.adapter.viewholder
 
 import com.yapp.itemfinder.domain.model.AddItemAdditional
 import com.yapp.itemfinder.feature.common.databinding.AddItemAdditionalBinding
+import com.yapp.itemfinder.feature.common.datalist.adapter.DataViewHolder
 import com.yapp.itemfinder.feature.common.extension.gone
 import com.yapp.itemfinder.feature.common.extension.visible
