+++ feature/common/src/main/java/com/yapp/itemfinder/feature/common/datalist/adapter/viewholder/AddItemPurchaseDateViewHolder.kt
@@ -1,7 +1,8 @@
-package com.yapp.itemfinder.feature.common.datalist.adapter
+package com.yapp.itemfinder.feature.common.datalist.adapter.viewholder
 
 import com.yapp.itemfinder.domain.model.AddItemPurchaseDate
 import com.yapp.itemfinder.feature.common.databinding.AddItemPurchaseDateBinding
+import com.yapp.itemfinder.feature.common.datalist.adapter.DataViewHolder
 
 class AddItemPurchaseDateViewHolder(
     val binding: AddItemPurchaseDateBinding
