+++ feature/common/src/main/java/com/yapp/itemfinder/feature/common/datalist/adapter/viewholder/AddItemSelectSpaceViewHolder.kt
@@ -1,7 +1,8 @@
-package com.yapp.itemfinder.feature.common.datalist.adapter
+package com.yapp.itemfinder.feature.common.datalist.adapter.viewholder
 
 import com.yapp.itemfinder.domain.model.AddItemSelectSpaceEntity
 import com.yapp.itemfinder.feature.common.databinding.ViewholderAddItemSelectSpaceBinding
+import com.yapp.itemfinder.feature.common.datalist.adapter.DataViewHolder
 
 class AddItemSelectSpaceViewHolder(
     private val binding: ViewholderAddItemSelectSpaceBinding
