+++ feature/common/src/main/java/com/yapp/itemfinder/feature/common/datalist/adapter/viewholder/CategoryViewHolder.kt
@@ -1,7 +1,8 @@
-package com.yapp.itemfinder.feature.common.datalist.adapter
+package com.yapp.itemfinder.feature.common.datalist.adapter.viewholder
 
 import com.yapp.itemfinder.domain.model.Category
 import com.yapp.itemfinder.feature.common.databinding.ViewholderStorageBinding
+import com.yapp.itemfinder.feature.common.datalist.adapter.DataViewHolder
 
 class CategoryViewHolder(binding: ViewholderStorageBinding) : DataViewHolder<Category>(binding) {
     override fun reset() {
